+++ Cargo.toml
@@ -15,6 +15,7 @@ spin = "0.9"
 modular-bitfield = "0.11"
 noto-sans-mono-bitmap = "0.2"
 conquer-once = { version = "0.4", default-features = false }
+fzproc_macros = { path = "src/fzboot/proc_macros" }
 
 [features]
 alloc = []
