--- src/fzboot/main/src/interrupts/mod.rs
@@ -1,1 +1,0 @@
-pub mod handlers;
