+++ src/fzboot/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "alloc")]
+pub mod irq;
+pub mod time;
