+++ src/lib.rs
@@ -1,21 +1,18 @@
 #![feature(allow_internal_unstable)]
+#![feature(proc_macro_hygiene)]
+#![feature(naked_functions)]
 #![no_std]
 #[macro_use]
 
 pub mod video;
 pub mod bios;
 pub mod fs;
+pub mod fzboot;
 pub mod io;
 pub mod mem;
-pub mod time;
 pub mod x86;
 
-
-#[cfg(feature = "alloc")]
-pub mod idt;
-
-pub mod int;
-
+pub use crate::fzboot::*;
 pub use numtoa;
 
 #[cfg(feature = "alloc")]
